+++ server/server.go
@@ -221,7 +221,7 @@ func (s *Server) startServer() error {
 	s.member.MemberInfo(s.cfg, s.Name(), s.rootPath)
 
 	s.idAllocator = id.NewAllocatorImpl(s.client, s.rootPath, s.member.MemberValue())
-	s.tso = tso.NewTimestampOracle(s.client, s.rootPath, s.member.MemberValue(), s.cfg.TsoSaveInterval.Duration, s.scheduleOpt)
+	s.tso = tso.NewTimestampOracle(s.client, s.rootPath, s.member.MemberValue(), s.cfg.TsoSaveInterval.Duration, func() time.Duration { return s.scheduleOpt.LoadPDServerConfig().MaxResetTSGap })
 	kvBase := kv.NewEtcdKVBase(s.client, s.rootPath)
 	path := filepath.Join(s.cfg.DataDir, "region-meta")
 	regionStorage, err := core.NewRegionStorage(path)
